docker/hasura-postgres-schema-sync-test/requirements.txt
pytest==7.1.2
requests==2.28.1
mock==4.0.3
python-dateutil==2.8.2
